files stack table updated for experiment processed files
